LiveLyricsPage 컴포넌트에 가사 미리보기 및 자막 미리보기 기능 추가: 가사 상태 계산 로직 개선, 활성 인덱스 추적 방식 수정, 불필요한 코드 제거로 성능 최적화.
